PLANET-7213 Refactor block files location
Ref. https://jira.greenpeace.org/browse/PLANET-7213
