+++ assets/src/block-editor/Sidebar/getSidebarFunctions.js
