+++ assets/src/block-editor/SidebarFields/ImageSidebarField.js
