+++ webpack.config.js
@@ -31,8 +31,8 @@ module.exports = {
     media_archive_editor_view: './assets/src/js/media_archive_editor_view.js',
     "lite-yt-embed": './node_modules/lite-youtube-embed/src/lite-yt-embed.js',
     menu_editor: './assets/src/js/menu_editor.js',
-    frontendIndex: './assets/src/blocks/frontendIndex.js',
-    editorIndex: './assets/src/blocks/editorIndex.js',
+    frontendIndex: './assets/src/frontendIndex.js',
+    editorIndex: './assets/src/editorIndex.js',
     GalleryScript: './assets/src/blocks/Gallery/GalleryScript.js',
     GalleryEditorScript: './assets/src/blocks/Gallery/GalleryEditorScript.js',
     GalleryStyle: './assets/src/scss/blocks/Gallery/GalleryStyle.scss',
